Read serial ouput from EZ0
This will use the serial output from the EZ0 instead of
the pulse output.
The distance is output in whole inches, so the code
has been refactored to handle this.

This removes much of the noise we have seen in the distance
readings, as the noise were introduced when we were reading the pulses.

+++ ultrasonic/src/main.h
@@ -33,10 +33,10 @@ typedef struct Buffer {
 	bool wrapped;
 } Buffer;
 
-int getPosition2D(Position2D *position, float distances[], unsigned int length);
+int getPosition2D(Position2D *position, unsigned int distances[], unsigned int length);
 void getPosition3D(Position3D *position, float r1, float r2, float r3);
 void getLine(Line *line, Position2D positions[], unsigned int length);
-void getInput(float distances[], unsigned int length);
+void getInput(unsigned int distances[], unsigned int length);
 bool willCollide2D(Line *line);
 void panic();
 bool isMoving(Position2D positions[], unsigned int length);
